requirements.txt
Flask
gunicorn
pandas
scikit-learn
spacy
joblib
python-dotenv
pyyaml
